relationship magazine-publishing house is checked for writing
