+++ templates/post_new_magazine.html
@@ -14,6 +14,8 @@
          <input id="quantity_in_stock" name="quantity_in_stock" type="text"/>
          <label>Аннотация:</label>
          <input id="description" name="description" type="text"/>
+         <label>Издательство:</label>
+         <input id="publishing_house" name="publishing_house" type="text"/>
          <button id="catalog" name="catalog">Перейти в каталог</button>
          <input id="new_magazine" value="Новый журнал" type="submit"/>
          <button id="new_book" name="new_book">Новая книга</button>
